requirements.txt
aiosqlite==0.22.1
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
auto-py-to-exe==2.48.1
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Eel==0.18.2
fastapi==0.135.1
ecdsa==0.19.1
pycryptodome==3.23.0
base58==2.1.1
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.1
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pefile==2024.8.26
pybit==5.14.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.10.2
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websocket-client==1.9.0
zope.event==6.1
zope.interface==8.2
